ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.1",
    "fastapi>=0.135.2",
    "finnhub-python>=2.4.27",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "langchain-community>=0.3.15",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
